Refactor and simplify code, update data types, clean tests
Refactored `ContributionService` and `TerminatedEmployeeAndBeneficiaryReport` for improved readability and maintainability. Updated `MemberSlice` record to change `PsnSuffix` type from `short` to `int`. Modified `MockDataContextFactory` to allow `SaveChangesAsync` without exceptions. Cleaned up `TerminatedEmployeeAndBeneficiaryTests` by removing unnecessary `using` directives and outdated test methods.

+++ src/services/src/Demoulas.ProfitSharing.Services/ServiceDto/MemberSlice.cs
@@ -6,7 +6,7 @@
 /// </summary>
 internal sealed record MemberSlice
 {
-    public short PsnSuffix { get; init; }
+    public int PsnSuffix { get; init; }
     public required int EmployeeId { get; init; }
     public long Ssn { get; init; }
     public decimal HoursCurrentYear { get; init; }
